Post-review amendments to the versioned CAS (lib-data-store-state-redis 1.2.1) [release]
Amendments from the post-merge review of #107:

- Refuse a versioned replace against a foreign frame-looking head whose
  digit run overruns the inspected head window (22+ digits): previously
  it fell through to version 0 like an unframed entry, so a version-0
  witness could overwrite it. Store-written frames (at most 26 bytes)
  are unaffected. Covered on both providers and at store level.
- Make VersionedStateStore serialize/deserialize final: an override
  dropping the frame would silently turn every CAS into a blind
  version-0 write.
- Restore the plain-store coverage narrowed by #107: AbstractStateStoreTest
  exercises AbstractStateStore again; versioned features moved to the
  new VersionedStateStoreTest.
- Document that no write returns the stamped version - a successful
  replaceIf invalidates the caller's own witness too, re-read before
  chaining - and scope the README retry comment to every cause of a
  false return.
