+++ lib-data-store-state-redis/changelog.txt
@@ -1,5 +1,21 @@
 # lib-data-store-state-redis changelog
 
+1.2.1 - 3 Aug 2026
+- Refuse a versioned replace against a foreign frame-looking head whose digit run
+  overruns the inspected head window (22+ digits): previously it fell through to
+  version 0 like an unframed entry, so a version-0 witness could overwrite it -
+  now no witness ever matches it, consistent with a digit run that overflows a
+  long within the window; store-written frames (at most 26 bytes) are unaffected
+- Make VersionedStateStore serialize/deserialize final: an override dropping the
+  version frame would silently turn every compare-and-swap into a blind version-0
+  write - customize the payload through the encoding strategy instead
+- Restore the plain-store test coverage narrowed by #107: AbstractStateStoreTest
+  exercises AbstractStateStore again and the versioned features moved to the new
+  VersionedStateStoreTest
+- Document that no write returns the stamped version: a successful replaceIf
+  invalidates the caller's own witness too - re-read before chaining another
+  replace; scope the README retry comment to every cause of a false return
+
 1.2.0 - 3 Aug 2026
 - Add VersionAware interface: a value carries its own optimistic-concurrency version -
   version() accessor plus withVersion(long) copy method (#107)
