+++ lib-data-store-state-redis/src/main/groovy/io/seqera/data/store/state/impl/RedisStateProvider.groovy
@@ -89,16 +89,17 @@ class RedisStateProvider implements StateProvider<String,String>, VersionProvide
      *
      * The digits are compared literally against the canonical decimal form of the
      * expected version - the only form a store-written frame can carry - so foreign
-     * data whose head merely resembles a frame never matches and is left untouched,
-     * as long as its digit run fits the 28-byte peek; a longer run (impossible for a
-     * store-written frame) leaves no terminator in the head and counts as version 0,
-     * same as an unframed value. LocalStateProvider mirrors the same comparison.
+     * data whose head merely resembles a frame never matches and is left untouched:
+     * a digit run whose terminator falls outside the 28-byte peek (impossible for a
+     * store-written frame, whose head is at most 26 bytes) is refused outright rather
+     * than falling through to unframed. LocalStateProvider mirrors the same comparison.
      */
     static private final String REPLACE_IF = '''
         local head = redis.call('GETRANGE', KEYS[1], 0, 27)
         if head == '' then return 0 end
-        local ver = string.match(head, '^{"@v":(%d+)[,}]') or '0'
-        if ver ~= ARGV[1] then return 0 end
+        local ver = string.match(head, '^{"@v":(%d+)[,}]')
+        if not ver and string.match(head, '^{"@v":%d') then return 0 end
+        if (ver or '0') ~= ARGV[1] then return 0 end
         redis.call('SET', KEYS[1], ARGV[2], 'PX', ARGV[3])
         return 1
         '''
